backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-dotenv==1.0.1
pydantic==2.9.0
python-dateutil==2.9.0
pyap==0.3.2
email-validator==2.2.0